extracted tag database logic to its own repository class.
